Fix investment plans with negative values for larger amounts
Issue: https://forum.portfolio-performance.info/t/sparplan-erzeugt-negative-stuecke-und-kurse/13952

+++ name.abuchen.portfolio/src/name/abuchen/portfolio/model/InvestmentPlan.java
@@ -435,8 +435,7 @@ public class InvestmentPlan implements Named, Adaptable, Attributable
                             converter.with(targetCurrencyCode).getRate(tDate, security.getCurrencyCode()).getValue());
             }
 
-            shares = Math.round((double) availableAmount * Values.Share.factor() * Values.Quote.factorToMoney()
-                / (double) price);
+            shares = Math.round((double) availableAmount * Values.Share.factor() * Values.Quote.factorToMoney() / price);
         }
 
         if (account != null)
